feat: Docker images tagged with Git tag (#56)
When a release is created it will now build a Docker image
tagged with that release's tag.

This change also sets the ERC image tag policy to IMMUTABLE
and removes the `latest` tag push from the GitHub workflow.  This
is because we're not using `latest` and it addresses a static
security analysis warning with the Terraform.