Swap out mysql for mariadb
This came up as an issue when using my cool new M1 Macbook Air:

- https://www.apple.com/ca/shop/buy-mac/macbook-air

Check it out, great battery life.

Anyway, there was an error booting up the mysql container, and I followed a few
of the threads in this stack overflow answer:

- https://stackoverflow.com/questions/65456814/docker-apple-silicon-m1-preview-mysql-no-matching-manifest-for-linux-arm64-v8

The best way forward seemed to be using mariadb as a drop-in replacement.
Works for me but I don't know if this works on everyone else's device.