Automate meta tag description for posts and pages
Basically, there are 3 cases:

- homepage: use site description
- posts pages: use first 150 characters of post
- categories pages, use category description

On the Canada Holidays site, I learned that there is a difference in
the amount of characters that will show up on Google vs. the amount
of characters that FB or Twitter will show, so you pretty much want
a shorter description for social media. But for, now we are good.