Add ref to Notice container and focus it on submit
Whether we fail or succeed, we always want to move focus up to the
Notice that comes through.

According to the W3C:
- The error container must be present in the DOM on page load for the error message to be spoken by most screen readers.

And then we can use a ref on that element as long as it's a basic HTML
element, not another functional component:
- You can, however, use the ref attribute inside a function component as long as you refer to a DOM element or a class component

So it all works, phew.

Sources:
- W3C: https://www.w3.org/TR/WCAG20-TECHS/ARIA19.html
- React docs: https://reactjs.org/docs/refs-and-the-dom.html#refs-and-function-components