Sanitize new usernames the same as email addresses
We are assigning the "email" to the username, but seeing different results.
Plus signs (+) and other special characters that are valid in emails are
being stripped from usernames, leading to potential for confusion down
the road.

This PR adds a filter to sanitize the username the same way we do emails.