Translate error messages
Also restructure error responses coming back from server.

I had made the unfortunate decision to return an object which included
"errors" as a sub-key of a key called "errors", so then it was really
confusing to actually find the errors.

Replaced with "errors.message", which is a string now and not an array.

We would only ever show the first error anyway, so no point of having
lots of different error messages.