Hide superadmins from user list
This action modifies the SQL to exlcude usernames that belong to superadmins.

This wasn't my first choice, but unfortunately the WP_User_Query
object doesn't want to allow filtering or sorting by usernames, only
IDs, and the way that superadmins are stored in the WordPress database
is sufficiently weird (they don't have a "superadmin" role).

Anyway, I tried a bunch of different things but this was the one
that worked.

Note that if someone who is a non-superadmin is able to see the
"network" users, they will also have the superadmins hidden, but I
don't believe this is possible.

Sources:
- Stackoverflow answer that almost helped: https://stackoverflow.com/questions/51667858/hide-users-with-a-specific-role-from-users-list-on-wordpress-admin-dashboard
- Imploding for SQL: https://stackoverflow.com/questions/10634058/wordpress-prepared-statement-with-in-condition
- https://developer.wordpress.org/reference/classes/wp_user_query/