Make TrackLogins more singleton-y
We ended up duplicating all the hooks before by using "new" two
times in the codebase for this class.

Rewrote it as kind of a singleton so that the "register" method will
create an instance set up the hooks, while getting a "new" class will
give you access to the methods you want to use.

Also added an extra param to the "getUserLogins" method so that
we can ignore my fake user agents.