Hooks to prevent new users being created on a new network site.
The code that creates a new network site is here: https://github.com/WordPress/WordPress/blob/a207dadc10e88208a82e6b94f204b0122df97d95/wp-admin/network/site-new.php#L36

It's a mix of functions and HTML code, so it's hard to extract anything in isolation without replacing the whole page. There are **3** hooks on this page, and no filters:

1. _Before_ adding a new user
2. _After_ adding a new user
3. _Before_ rendering the "submit" button (but _after_ the rest of the form has already been rendered)

- Using the 1st hook, we can prevent new users from being added (by using `wp_die` — returning a `WP_Error` is ignored).
- Using the 3rd hook, we can add more HTML to the page. In this case, I used CSS to hide the previous message about creating new users and I added new message saying that only Super Admins are accepted.

Unfortunately, there is no way to override parts of this form (for example, removing the "Admin Email" field and replacing it with a `<select>`), or the underlying logic for checking users (we can only stop sites being created for new users, we can't check the role of existing users).