Remove "home" class, replace with "index"
By sheer conincidence, the "WET" theme has some special styling rules
for some elements on the "home" page, by which they mean only Canada.ca.

WordPress's "body_class()" function sticks "home" in the list of headers
as well, which we don't want, since we're not trying to be Canada.ca.

I've replaced "home" with "homepage", so we can still use styling on this
page if we want, and I've added one CSS rule to keep our headers small
(otherwise they are enormous).