Allow GC Admins and Editors to edit Menus
The wordpress capability that governs this is called "edit_theme_options",
but it also allows users to edit the theme itself and use that
"Customize" menu, which we don't want to allow.

Therefore, there are a few things in this commit:
- Add "edit_theme_options" capabilites for GC Admin and GC Editor
- Add "Menus" link in admin menu ("Appearance" remains hidden)
- Explicitly return wp_die error if a non-superadmin is trying to
  edit themes or reach the "customize" menu

The "wp_die" page returns the same message they currently get.