Fix side nav styles with really specific CSS
It looks like the CSS file got updated online and so we need to
be more specific than the styles they are using in order to get
our side nav to work.

FAQ:

Why are we using this wordpress classes if they conflict with
our own CSS?

Basically, it's because I want to keep the markup as similar as possible
to what you would get using Gutenberg. It's kind of a weak rationale,
but there it is.