Add drop-down and mobile options for main menu
The drop-down menu markup was taken from CSS tricks
- https://css-tricks.com/solved-with-css-dropdown-menus/

It also suggests using an aria-label, so I've added one.

On large screens, it opens when you hover or focus the parent element.
On small screens, we just list all the links, so it doesn't collapse.

The idea was that we don't want to create a drop-down that's just
a button, where you can't click the link anymore.

The mobile nav is an "open/close" pattern that I copied from Bootstrap.
- https://getbootstrap.com/docs/5.1/components/navbar/

Get the header HTML, insert a button before the <ul>, and then added
some CSS and JS to get the whole thing functional.
I've copied what bootstrap does with aria-roles, which is pretty minimal.

All sources:
- https://css-tricks.com/solved-with-css-dropdown-menus/
- https://getbootstrap.com/docs/5.1/components/navbar/
- https://github.com/wasinger/htmlpagedom
- https://stackoverflow.com/questions/51466367/append-html-to-a-tag-using-php-html-parser#comment89901301_51466367