Add new Cleanup class to add print stylesheet
Adding print styles to the site. I've kept them in their own file which is only applied when `media="print"`.

Just running through the list of changes:
- Keep the FIP fairly small
- Remove padding on breadcrumbs, otherwise they take up lots of space
- Force white text to be black (otherwise you can't see it)
- Hide top and side menus
- Remove padding on `wp-block-columns` otherwise they are too spread out
- Hide images without alt text
- Reduce all headings and text sizes by 20%

We can make updates to this later if we find it's not convenient for our users, but trying locally with some of the pages that we have, it's a definite improvement.