Add language link to the login page
Some considerations:

WP can give you your current URL, but not an alternate URL.

The WPML docs aren't too much help here either. It seems like
the login page is a special case for them. Their documentation
shows how to translate the strings, but not how to programatically
get the login page URL for other langs.

Luckily, it's fairly trivial to do this since we know what our two
languages are.

Added a language switching link under the header, so that it's one
of the first things you come across.
Copied the styles from the links underneath so that it has the same
colours and padding.

Sources:
- https://wpml.org/documentation/getting-started-guide/translating-wordpress-login-and-registration-pages/
- https://wpml.org/documentation/support/wpml-coding-api/