Add language switcher to the top #wpadminbar
Unlike the language switcher on the public facing site, this one
changes the language in the wordpress interface.

Getting this to work was fairly tricky to do because the
admin bar is for links, not for putting random HTML into, but in
the end it seems to have worked out okay.

The other plugin created an extra script file and made it an Ajax
request, but I didn't want to do that.