Remove yubikeys from a user's profile
Don't show the last column in the table for Two-Factor Options, which
allows users to add Yubikeys.

If you already have a Yubikey assigned, you can still select another
option.

What you won't be able to do is remove an existing Yubikey, but for
all practical purposes, this shouldn't matter.