Stop changing default language
I thought this would change the WPML language in the backend, but
it changes the default language for the site itself.

This is actually a very disruptive action.

If previously our site was
- (EN homepage) articles.ca/home
- (FR homepage) articles.ca/fr/home

Changing this setting changes the URLs:
- (EN homepage) articles.ca/en/home
- (FR homepage) artciles.ca/home

So we don't want to change this when the admins change their
language in the backend, it's a terrible idea.