Use box-shadow instead of outline for focus styles on footer links
Fixes a weird CSS bug that happens where having a focus outline
seems to break the column css in Safari.

@amazingphilippe pointed it out to us.
Basically tabbing through the footer links in safari will cause
the links to change size and push each other between the different
columns.

Outlines are not supposed to affect the size of elements, so this is a bug to do with Safari.