Add a URL preview under the site title
Added a URL preview widget that will give people a sense of what
their URL will look like when they are submitting a request for a title.

Asked our accessibility tutor what we should do with a message that
updates itself (but isn't an error!).

She suggested making it an aria-live region with a "polite" attribute,
and also using a "described by" attribute on the input, so that it
will at least signal that there's a longer explanation.

Tried this out in voiceover and it seemed to work, so let's ship it.

Sources:
- https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/ARIA_Live_Regions
- https://ricardometring.com/javascript-replace-special-characters