Hide "Canada.ca" menu + unassign custom menu if exists (#638)
* Add site setting to hide the Canada.ca menu

Same as we do for the search bar or the breadcrumbs, add a class
and then hide it.

Keeping the blue line for now. I think it adds scructure to the page.

* Add a "warning" message if a custom menu exists

* Add logic for setting and unsetting menus

Essentially, we have 3 different states we can be in:

- Canada.ca top menu
- We have our own top menu that we set
- We don't want a top menu

To simplify things, we are going to assume

- someone who clicks option (3) wants to get rid of any top menu they have assigned (whether 1 OR 2)
- someone who assigns a top menu (option 2) wants to unset option (3)

What we want to avoid:

- someone builds their own menu but is not able to ever see it because they previously selected "hide canada.ca top menu" (in this case, it keeps getting programatically unassigned)
- someone sets "hide canada.ca top menu" and then it still still shows their custom menu that they don't want

* Use hook that happens later in the WP lifecycle

Previously, using the "init" hook, the menu location hadn't been
set yet, so the condition would fail.