Add checklists permission for admins and gceditors (#694)
Adds `manage_checklists` permissions to administrator and gceditor roles
when checklist plugin is active, removes them when the plugin is deactivated.

Took me a while to figure out because we also filter menu items. So
even when the permission was working, the menu page was not showing up.
Eventually figured it out, but ended up going down a few dead ends.