Make sure the yoast panel is closed by default
This was kind of a weird one.

There's a gutenberg API for closing elements, which most metaboxes and
panels reqpect, but the Yoast Panel doesn't.
It does set the Yoast Panel's "hidden" attribute correctly though,
so I am still including it to make sure that the data is consistent.

We can use a JS selector to press the "hide" button, but we need to
wait for the button to show up before it actually hides itself.
Added a function that polls for the button (for 4 seconds), and will hide
the panel if it sees the button appear.

Sources:
- https://wordpress.stackexchange.com/questions/367619/initial-open-toggle-plugindocumentsettingpanel-panel-in-document-setting-block-e
- https://developer.wordpress.org/block-editor/reference-guides/data/data-core-edit-post/#iseditorpanelopened
- https://developer.wordpress.org/block-editor/reference-guides/slotfills/plugin-document-setting-panel/#accessing-a-panel-programmatically
- https://stackoverflow.com/a/29754070