Allow GC Admins to embed unsafe HTML content (eg, iframes)
Unlike our other capabilites, which we can add to roles during the
setup, `unfiltered_html` is treated differently.

> User roles other than Super Admin cannot be assigned the unfiltered_html capability in WordPress Multisite. You can set it, but then WordPress disables the capability after the fact.

Because of this, we have to add the 'unfiltered_html' capability
in real time by overriding the 'map_meta_cap' filter.

Adding unfiltered_html to the roles array does not work, but this does.

Source:
- https://smartslider.helpscoutdocs.com/article/1983-how-to-give-access-to-smart-slider-for-non-admin-users#multisite