Return nav menus independently of each other via the API
There is a 'helpful' WPML function that will try and always return
the nav_menu in the current language of the site.

This leads to the same menu being returned twice when you ask for
"all menus" via the API, which is NOT what we want.

Thought of adding this only to the Redirector theme, but since the
API returns menus the same way in both themes, I can't think of a case
when we would want the current behaviour.