Make the contact form into a 2-step form
Since we started to get spammed on the Contact form, we are splitting it into 2 pages. The result will be that's easy for humans to fill out but hard for bots.

The majority of this code is brought over from the "Request a site" form, which is already broken into 2 pages.

It's not the most beautiful code in the world but it's been working okay for what we need it for.

- Updated the "group" divs with an id so that clicking the error messages takes you to that question
- Updated the ids of the optional fields because it seemed too confusing to have a div wrapping the optional fields with an id of `optional-usage` and the field itself with an id of `usage-optional`
- Fixed a pretty minor axe problem where some of the description ids on the contact form were duplicated