Form values are preserved when page returned with an error
Previously, the contact form wouldn't save your values if you
made an error and were returned to the same page.

The Request a site form would, but it didn't work for optional
values and it didn't work in French.

This commit fixes that bug for both forms.
The easiest way to test this is to leave one of the checkbox
fields unchecked, and try to submit with everything else filled out.