Update the tests for the inputs so that we check for IDs instead of values
We were using the values to check if someone's old answers should be
pre-populated in the form.
However, once we switch to French, the "value" switches to French as well,
so you end up losing your answer.

Matching against the ID instead solves this problem, because the "id"
is the actual value that is saved whereas the "value" string is
just used as the label.