Fix the hover closing state for our flyout menus
How the menu used to work was really annoying. You had to exit the
submenu specifically to get the menu to close itself (after a 1 second delay).

This meant that if you triggered the button and then just left by moving your mouse
up or sideways, that the menu would stay open forever.

Changed the "mouseout" selector so that it applies to everything inside
the "has-submenu" link element, which means that it will trigger the
delay to close the menu no matter where you leave it from.

I've also shortened the delay to half a second, based on advice in an
article I was reading:

> We need to make sure that the menu doesn’t open and doesn’t close too early. To achieve that, we introduce a delay, usually around 0.5 seconds.

Source:

- https://www.smashingmagazine.com/2021/05/frustrating-design-patterns-mega-dropdown-hover-menus/