Fill out new POST route for assigning a translation to a post/page
There's a helpful WPML action where you can reassign a post's "trid"
value, or set it to `false`, which WPML figures out how to increment it,
but other than that, you're on your own.

This route is deliberately permissive.

There are 4 cases here:

1. New post (no trns) -> other post (no trns)
2. New post (w/ trns) -> other post (no trns)
3. New post (no trns) -> other post (w/ trns)
4. New post (w/ trns) -> other post (w/ trns)

If we prevented scenarios 2, 3, and 4, we would be in a position where you
would have no way to reassign a translation once you assigned one, other than
deleting your post and starting over.

In the scenarios where a translation is already assigned, the logic will
create orphans of the previously-linked posts and make sure the current one is set.

Eg,

Say we have post 1 (EN), post 2 (FR), and post 3 (EN).

If post 1 and 2 are linked, but we submit a request to link 3 -> 2, then
post 1 will become an orphan and 3 -> 2 will be linked.

There are also plenty of ways this can throw back errors, I think I've thought
of all the failure cases.