Update the post_meta if there is one when incrementing the trid
The post_meta was holding on to an old value for the trid, based on the
original one the post started out with.

Since it was updating the post on save, it meant that it would revert
posts to the old trid they had when they were first initialized,
which would mean you would have to set the transation again.

If we update the post_meta (where it exists) to a new post_meta, all
of our problems are solved.