Fix all the translations bugs (#1021)
* Remove the WPML metabox to kind of improve things

The WPML metabox panel is removed using "removeEditorPanel" in sidebar.js, but that
just hides the metabox using CSS, it's not actually removed from the DOM.

The WPML metabox loads with a pre-populated select box with the name of the post that is
set to the current translation when the page loads. This value is then submitted as part of
a general "POST" request when the page/post is updated. WPML looks for this value and,
when it finds it, it uses it as the translation.

What this means is that on some posts, our new translation ID is sent, the translation is updated,
but then half a second later the "icl_translation_of" value is received by WPML, which resets the translation to its prior value.

Our solution to this is to remove the WPML metabox from the DOM if we find it.
It would be better to remove the WPML panel completely using a WordPress API but this doesn't seem to be possible.

* Update the post_meta if there is one when incrementing the trid

The post_meta was holding on to an old value for the trid, based on the
original one the post started out with.

Since it was updating the post on save, it meant that it would revert
posts to the old trid they had when they were first initialized,
which would mean you would have to set the transation again.

If we update the post_meta (where it exists) to a new post_meta, all
of our problems are solved.