fix: lower WAF rate limit to 2,000 requests
Update the WAF rate limit to only allow 2,000 requests from a single
IP in a 5 minute period.  This is in response to the fuzzing attacks
we've seen that have caused problems, but been below the previous
limit of 10,000.

Also re-enable the WAF in Staging now that the pentesting exercise
is complete.