Bump up lockout times to a year
Previously, if someone didn't log in for 90 days, we would lock
their account. This was a security feature to just be more proactive around
dormant accounts.

However, it seems we have some sleepy users and there are a number
of legitimate users who don't log in for longer than 90 days and then
need to be unlocked.

Bryan has requested the lockout time be 1 year, so that's what we're
going to go with.

Source:
- https://github.com/cds-snc/gc-articles-issues/issues/476