Allow self-hosted files to be shown on GC Articles
Previously, our CSP rules would ban any self-hosted PDFs from being
displayed (they had to come from Canada.ca), which seems overly-zealous.

This change allows PDFs to be uploaded to GC articles and then displayed on the frontend,
as well as safelisting a french WET script that we missed and a WPML script.