Copy template text to clipboard
Kind of a weird way to do it, but at least it keeps all the text
in one place, so that we don't go out of sync.

The idea is that we specify all the strings in the PHP layer, and
then we can both echo it out to our template as well as send it to
the JS layer as a localized variable.