fix: invalidate entire subsite cache (#1285)
Update the `CDS base` plugin `c3-cloudfront-clear-cache` 
filter so that cache invalidation requests cause the entire 
subsite's CloudFront cache to be cleared.

This is to work around a bug in the plugin that is preventing 
any cache invalidations from running.  There are several open 
issues and PRs on the plugin to fix the issue but they haven't 
been merged or triaged.