fix: suppress intermittent proxy getaddrinfo errors
Update the CloudWatch error alarms so that occasional failures to
get the RDS proxy address info do not trigger a Slack post.

These will now only be alerted if there are more than 5 in a short
period of time.

Also fix a TF flip-flop with the custom RDS parameter group.