Infrastructure release 1.5.12 (#2175)
Deploy WAF IP blocklist to production. This change will automatically manage 
a list of IPs that should be temporarily blocked from accessing the service 
based on their previous requests.

The firewall rule is currently set to COUNT mode so we can evaluate if this 
works as expected before we switch it to start blocking traffic.